Feature: Update Clear button
